+++ src/Views/PostItem/PostItem.tsx
@@ -1,42 +1,18 @@
+import { useState } from "react";
 import EllipsisSvg from "../../Assets/SVG/EllipsisSvg";
 import UserPhoto from "../../sass/styled-components/UserPhoto";
+import { timeFormat } from "../../utills/function/function";
 import InteractionBar from "../InteractionBar/InteractionBar";
 import { dataType } from "../PostList/PostList";
 import "./PostItem.scss";
 import sample from "/Users/blanc/Documents/Project/sns/src/Assets/sample.png";
-// import sample from "/Users/blanc/Documents/Project/sns/src/Assets/vertical_sample.jpeg";
+
 // type
 interface dataTypeProps {
   data: dataType;
 }
 const PostItem = ({ data }: dataTypeProps) => {
-  function timeForToday(value: string) {
-    const today = new Date();
-    const timeValue = new Date(value);
-
-    const betweenTime = Math.floor(
-      (today.getTime() - timeValue.getTime()) / 1000 / 60
-    );
-    if (betweenTime < 1) return "방금 전";
-    if (betweenTime < 60) {
-      return `${betweenTime}분 전`;
-    }
-
-    const betweenTimeHour = Math.floor(betweenTime / 60);
-    if (betweenTimeHour < 24) {
-      return `${betweenTimeHour}시간 전`;
-    }
-
-    const betweenTimeDay = Math.floor(betweenTime / 60 / 24);
-    if (betweenTimeDay < 365) {
-      if (betweenTimeDay < 2) {
-        return `어제`;
-      }
-      return `${betweenTimeDay}일 전`;
-    }
-
-    return `${Math.floor(betweenTimeDay / 365)}년 전`;
-  }
+  const [likes, setLikes] = useState(data.likes);
 
   console.log("postItem", data);
   return (
@@ -55,8 +31,8 @@ const PostItem = ({ data }: dataTypeProps) => {
           <img src={sample} alt="sample" />
         </div>
         <div className="PostItem__bottom">
-          <InteractionBar />
-          <div className="PostItem__bottom--likes">{data.likes} likes</div>
+          <InteractionBar likes={likes} setLikes={setLikes} />
+          <div className="PostItem__bottom--likes">{likes} likes</div>
           <div className="PostItem__bottom--content">
             <span>{data.nickName}</span>
             <span>{data.text}</span>
@@ -65,7 +41,7 @@ const PostItem = ({ data }: dataTypeProps) => {
             View all {data.commentsNumber} comments
           </div>
           <div className="PostItem__bottom--posted-at">
-            {timeForToday(data.postedAt)}
+            {timeFormat(data.postedAt)}
           </div>
         </div>
       </section>
