+++ src/Views/PostList/PostList.tsx
@@ -1,4 +1,6 @@
 import { useState } from "react";
+import { useRecoilState } from "recoil";
+import { datasState } from "../../recoil/snsState";
 import PostItem from "../PostItem/PostItem";
 import "./PostList.scss";
 
@@ -13,30 +15,8 @@ export interface dataType {
   userImg: string;
 }
 
-export const initialdatas = [
-  {
-    postId: 1,
-    postImg: "/Users/blanc/Documents/Project/sns/src/Assets/sample.png",
-    likes: 9,
-    nickName: "blanc",
-    text: "test입니다.",
-    commentsNumber: 4,
-    postedAt: "Sat Jan 21 2023 17:29:28",
-    userImg: "/Users/blanc/Documents/Project/sns/src/Assets/blanc.jpeg",
-  },
-  {
-    postId: 2,
-    postImg: "/Users/blanc/Documents/Project/sns/src/Assets/sample.png",
-    likes: 9,
-    nickName: "blanc",
-    text: "test입니다.",
-    commentsNumber: 4,
-    postedAt: "Sat Jan 20 2023 13:39:28",
-    userImg: "/Users/blanc/Documents/Project/sns/src/Assets/blanc.jpeg",
-  },
-];
 const PostList = () => {
-  const [datas, setDatas] = useState<dataType[]>(initialdatas);
+  const [datas, setDatas] = useRecoilState<dataType[]>(datasState);
 
   console.log("postList", datas);
   return (
