+++ src/index.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { RecoilRoot } from "recoil";
 import App from "./App";
 import "./index.scss";
 import reportWebVitals from "./reportWebVitals";
@@ -8,9 +9,9 @@ const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
-  <>
+  <RecoilRoot>
     <App />
-  </>
+  </RecoilRoot>
 );
 
 reportWebVitals();
