+++ src/utills/function/function.ts
@@ -1,0 +1,27 @@
+export function timeFormat(value: string) {
+  const today = new Date();
+  const timeValue = new Date(value);
+
+  const betweenTime = Math.floor(
+    (today.getTime() - timeValue.getTime()) / 1000 / 60
+  );
+  if (betweenTime < 1) return "방금 전";
+  if (betweenTime < 60) {
+    return `${betweenTime}분 전`;
+  }
+
+  const betweenTimeHour = Math.floor(betweenTime / 60);
+  if (betweenTimeHour < 24) {
+    return `${betweenTimeHour}시간 전`;
+  }
+
+  const betweenTimeDay = Math.floor(betweenTime / 60 / 24);
+  if (betweenTimeDay < 365) {
+    if (betweenTimeDay < 2) {
+      return `어제`;
+    }
+    return `${betweenTimeDay}일 전`;
+  }
+
+  return `${Math.floor(betweenTimeDay / 365)}년 전`;
+}
